feat: permissions for random tables

+++ routers/random_table_router.ts
@@ -4,15 +4,21 @@ import { jsonArrayFrom } from "kysely/helpers/postgres";
 import { DB } from "kysely-codegen";
 
 import { db } from "../database/db";
+import { getHasEntityPermission } from "../database/queries";
 import { EntitiesWithChildren } from "../database/types";
 import { EntityListSchema } from "../database/validation";
 import { InsertRandomTableSchema, ReadRandomTableSchema, UpdateRandomTableSchema } from "../database/validation/random_tables";
 import { MessageEnum } from "../enums/requestEnums";
-import { beforeRoleHandler } from "../handlers";
+import { beforeRoleHandler, noRoleAccessErrorHandler } from "../handlers";
 import { PermissionDecorationType, ResponseSchema, ResponseWithDataSchema } from "../types/requestTypes";
 import { constructFilter } from "../utils/filterConstructor";
 import { constructOrdering } from "../utils/orderByConstructor";
-import { GetEntityChildren, GetParents, GetRelationsForUpdating } from "../utils/relationalQueryHelpers";
+import {
+  CreateEntityPermissions,
+  GetEntityChildren,
+  GetParents,
+  GetRelationsForUpdating,
+} from "../utils/relationalQueryHelpers";
 import { getEntityWithOwnerId } from "../utils/transform";
 
 export function random_table_router(app: Elysia) {
@@ -42,6 +48,9 @@ export function random_table_router(app: Elysia) {
                   const withParentId = random_table_options.map((opt) => ({ ...opt.data, parent_id: id }));
                   await tx.insertInto("random_table_options").values(withParentId).execute();
                 }
+              }
+              if (body.permissions?.length) {
+                await CreateEntityPermissions(tx, id, "random_table_permissions", body.permissions);
               }
             });
             return { message: `Random table ${MessageEnum.successfully_created}`, ok: true, role_access: true };
@@ -138,53 +147,59 @@ export function random_table_router(app: Elysia) {
         )
         .post(
           "/update/:id",
-          async ({ params, body }) => {
-            await db.transaction().execute(async (tx) => {
-              await tx.updateTable("random_tables").where("id", "=", params.id).set(body.data).executeTakeFirstOrThrow();
+          async ({ params, body, permissions }) => {
+            const permissionCheck = await getHasEntityPermission("random_tables", params.id, permissions);
+            if (permissionCheck) {
+              await db.transaction().execute(async (tx) => {
+                await tx.updateTable("random_tables").where("id", "=", params.id).set(body.data).executeTakeFirstOrThrow();
 
-              if (body?.relations) {
-                if (body.relations?.random_table_options) {
-                  const { random_table_options } = body.relations;
-                  const existingCharacterFields = await tx
-                    .selectFrom("random_table_options")
-                    .select(["id"])
-                    .where("random_table_options.parent_id", "=", params.id)
-                    .execute();
-                  const existingIds = existingCharacterFields.map((field) => field.id);
-                  const [idsToRemove, itemsToAdd, itemsToUpdate] = GetRelationsForUpdating(
-                    existingIds,
-                    random_table_options.map((opt) => opt.data),
-                  );
-                  if (idsToRemove.length) {
-                    await tx.deleteFrom("random_table_options").where("id", "in", idsToRemove).execute();
-                  }
-                  if (itemsToAdd.length) {
-                    await tx
-                      .insertInto("random_table_options")
-                      .values(
-                        itemsToAdd.map((item) => ({
-                          parent_id: params.id,
-                          title: item.title,
-                          description: item.description,
-                        })),
-                      )
+                if (body?.relations) {
+                  if (body.relations?.random_table_options) {
+                    const { random_table_options } = body.relations;
+                    const existingCharacterFields = await tx
+                      .selectFrom("random_table_options")
+                      .select(["id"])
+                      .where("random_table_options.parent_id", "=", params.id)
                       .execute();
-                  }
-                  if (itemsToUpdate.length) {
-                    await Promise.all(
-                      itemsToUpdate.map(async (item) => {
-                        await tx
-                          .updateTable("random_table_options")
-                          .where("parent_id", "=", params.id)
-                          .where("id", "=", item.id)
-                          .set({ title: item.title, description: item.description })
-                          .execute();
-                      }),
+                    const existingIds = existingCharacterFields.map((field) => field.id);
+                    const [idsToRemove, itemsToAdd, itemsToUpdate] = GetRelationsForUpdating(
+                      existingIds,
+                      random_table_options.map((opt) => opt.data),
                     );
+                    if (idsToRemove.length) {
+                      await tx.deleteFrom("random_table_options").where("id", "in", idsToRemove).execute();
+                    }
+                    if (itemsToAdd.length) {
+                      await tx
+                        .insertInto("random_table_options")
+                        .values(
+                          itemsToAdd.map((item) => ({
+                            parent_id: params.id,
+                            title: item.title,
+                            description: item.description,
+                          })),
+                        )
+                        .execute();
+                    }
+                    if (itemsToUpdate.length) {
+                      await Promise.all(
+                        itemsToUpdate.map(async (item) => {
+                          await tx
+                            .updateTable("random_table_options")
+                            .where("parent_id", "=", params.id)
+                            .where("id", "=", item.id)
+                            .set({ title: item.title, description: item.description })
+                            .execute();
+                        }),
+                      );
+                    }
                   }
                 }
-              }
-            });
+              });
+            } else {
+              noRoleAccessErrorHandler();
+              return { message: "", ok: false, role_access: false };
+            }
 
             return { message: `Random table ${MessageEnum.successfully_updated}.`, ok: true, role_access: true };
           },
@@ -196,9 +211,15 @@ export function random_table_router(app: Elysia) {
         )
         .delete(
           "/:id",
-          async ({ params }) => {
-            await db.deleteFrom("random_tables").where("id", "=", params.id).execute();
-            return { message: `Random table ${MessageEnum.successfully_deleted}`, ok: true, role_access: true };
+          async ({ params, permissions }) => {
+            const permissionCheck = await getHasEntityPermission("random_tables", params.id, permissions);
+            if (permissionCheck) {
+              await db.deleteFrom("random_tables").where("id", "=", params.id).execute();
+              return { message: `Random table ${MessageEnum.successfully_deleted}`, ok: true, role_access: true };
+            } else {
+              noRoleAccessErrorHandler();
+              return { message: "", ok: false, role_access: false };
+            }
           },
           { response: ResponseSchema, beforeHandle: async (context) => beforeRoleHandler(context, "delete_random_tables") },
         ),
